Add task

+++ 0x10-variadic_functions/3-print_all.c
@@ -33,7 +33,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(p, int)), k = 1;
 				break;
 			case 'f':
-				printf("%f", va_arg(p, int)), k = 1;
+				printf("%f", va_arg(p, double)), k = 1;
 				break;
 			case 's':
 				str = va_arg(p, char *), k = 1;
